refactor: CarVO -> CarResponse로 역할에 맞게 수정
